+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-flag/components/add-comments/add-comments.component.ts
@@ -27,7 +27,8 @@ export class AddCommentsComponent implements OnInit {
   public ngOnInit(): void {
     this.addCommentsTitle = !this.isDisplayContextParameterExternal ?
       CaseFlagWizardStepTitle.ADD_FLAG_COMMENTS : CaseFlagWizardStepTitle.ADD_FLAG_COMMENTS_EXTERNAL_MODE;
-    this.addCommentsHint = AddCommentsStep.HINT_TEXT;
+    this.addCommentsHint = !this.isDisplayContextParameterExternal ?
+      AddCommentsStep.HINT_TEXT : AddCommentsStep.HINT_TEXT_EXTERNAL;
     this.addCommentsCharLimitInfo = AddCommentsStep.CHARACTER_LIMIT_INFO;
 
     if (!this.formGroup.get(this.flagCommentsControlName)) {
@@ -47,10 +48,12 @@ export class AddCommentsComponent implements OnInit {
     this.flagCommentsCharLimitErrorMessage = null;
     this.errorMessages = [];
     if (!this.optional && !this.formGroup.get(this.flagCommentsControlName).value) {
-      this.flagCommentsNotEnteredErrorMessage = AddCommentsErrorMessage.FLAG_COMMENTS_NOT_ENTERED;
+      this.flagCommentsNotEnteredErrorMessage = this.isDisplayContextParameterExternal
+        ? AddCommentsErrorMessage.FLAG_COMMENTS_NOT_ENTERED_EXTERNAL
+        : AddCommentsErrorMessage.FLAG_COMMENTS_NOT_ENTERED;
       this.errorMessages.push({
         title: '',
-        description: AddCommentsErrorMessage.FLAG_COMMENTS_NOT_ENTERED,
+        description: this.flagCommentsNotEnteredErrorMessage,
         fieldId: this.flagCommentsControlName
       });
     }