+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-flag/components/update-flag/update-flag.component.ts
@@ -79,11 +79,12 @@ export class UpdateFlagComponent implements OnInit {
     switch (this.displayContextParameter) {
       case CaseFlagDisplayContextParameter.UPDATE:
         if (flagDetail?.name) {
-          return `${CaseFlagWizardStepTitle.UPDATE_FLAG_TITLE} "${flagDetail.name}${flagDetail.subTypeValue ? `, ${flagDetail.subTypeValue}"` : '"'}`;
+          const subTypeValue = flagDetail.subTypeValue ? `, ${flagDetail.subTypeValue}` : ''
+          return `${CaseFlagWizardStepTitle.UPDATE_FLAG_TITLE} "${flagDetail.name}${subTypeValue}"`;
         }
         return `${CaseFlagWizardStepTitle.UPDATE_FLAG_TITLE}`;
       case CaseFlagDisplayContextParameter.UPDATE_EXTERNAL:
-        return CaseFlagWizardStepTitle.UPDATE_FLAG_TITLE_SUPPORT;
+        return CaseFlagWizardStepTitle.UPDATE_FLAG_TITLE_EXTERNAL;
       default:
         return CaseFlagWizardStepTitle.NONE;
     }
@@ -114,10 +115,10 @@ export class UpdateFlagComponent implements OnInit {
     // is no existing comment then one is not required for validation to pass
     const comment = this.formGroup.get(CaseFlagFormFields.COMMENTS).value;
     if (!comment && (this.flagDetail.flagComment || this.flagDetail.flagComment_cy)) {
-      this.updateFlagNotEnteredErrorMessage = UpdateFlagErrorMessage.FLAG_COMMENTS_NOT_ENTERED;
+      this.updateFlagNotEnteredErrorMessage = this.getUpdateFlagNotEnteredErrorMessage();
       this.errorMessages.push({
         title: '',
-        description: UpdateFlagErrorMessage.FLAG_COMMENTS_NOT_ENTERED,
+        description: this.updateFlagNotEnteredErrorMessage,
         fieldId: CaseFlagFormFields.COMMENTS
       });
     }
@@ -149,6 +150,17 @@ export class UpdateFlagComponent implements OnInit {
         description: UpdateFlagErrorMessage.STATUS_REASON_CHAR_LIMIT_EXCEEDED,
         fieldId: CaseFlagFormFields.STATUS_CHANGE_REASON
       });
+    }
+  }
+
+  private getUpdateFlagNotEnteredErrorMessage(): UpdateFlagErrorMessage {
+    switch(this.displayContextParameter) {
+      case CaseFlagDisplayContextParameter.UPDATE:
+        return UpdateFlagErrorMessage.FLAG_COMMENTS_NOT_ENTERED;
+      case CaseFlagDisplayContextParameter.UPDATE_EXTERNAL:
+        return UpdateFlagErrorMessage.FLAG_COMMENTS_NOT_ENTERED_EXTERNAL;
+      default:
+        return UpdateFlagErrorMessage.NONE;
     }
   }
 }