+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-flag/enums/case-flag-check-your-answers-step.enum.ts
@@ -1,0 +1,10 @@
+export enum CaseFlagCheckYourAnswersPageStep {
+  CASE_LEVEL_LOCATION = 'Case level',
+  ADD_FLAG_HEADER_TEXT = 'Add flag to',
+  ADD_FLAG_HEADER_TEXT_EXTERNAL = 'Add support to',
+  UPDATE_FLAG_HEADER_TEXT = 'Update flag for',
+  UPDATE_FLAG_HEADER_TEXT_EXTERNAL = 'Update support for',
+  FLAG_TYPE_HEADER_TEXT = 'Flag type',
+  FLAG_TYPE_HEADER_TEXT_EXTERNAL = 'Support type',
+  NONE = ''
+}