+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-flag/enums/update-flag-step.enum.ts
@@ -1,5 +1,5 @@
 export enum UpdateFlagStep {
   COMMENT_HINT_TEXT = 'Update the comments describing the user\'s support needs or flag description. Do not include any sensitive information such as personal details.',
   CHARACTER_LIMIT_INFO = 'You can enter up to 200 characters',
-  STATUS_HINT_TEXT = 'Describe reason for status change; if choosing \'Not approved\' provide name of person approving decision.'
+  STATUS_HINT_TEXT = 'Describe reason for status change.',
 }